requirement.txt

django==1.8.5
weixin-python